change long long to long

+++ more_malloc_free/2-calloc.c
@@ -34,7 +34,7 @@ int _strlen(char *s)
 void *_calloc(unsigned int nmemb, unsigned int size)
 {
 	char *ptr;
-	unsigned long long int i;
+	unsigned long int i;
 
 	if (nmemb == 0 || size == 0)
 	{
